On CustomGestureDetector : In the SharedPref, added the backgroundcolor after the swipe to register it.
On MainActivity :
Finalised the startAlarm function.
When we launch the app, the programm is now checking if a json is registered. (json of MoodHistory for the view history activity)
If there is one, we keep it.
If not, we create a new one.

On AlertReceiver :
Copied the code from the AlarmTest button from the mainActivity